+++ eslint.config.js
@@ -5,7 +5,11 @@ import eslintPluginPrettier from "eslint-plugin-prettier/recommended";
 
 /** @type {import('eslint').Linter.Config[]} */
 export default [
-  { languageOptions: { globals: { ...globals.browser, ...globals.node } } },
+  {
+    languageOptions: {
+      globals: { ...globals.browser, ...globals.node, $: "readonly" },
+    },
+  },
   pluginJs.configs.recommended,
   eslintPluginPrettier,
   eslintConfigPrettier,
