+++ src/util/login.js
@@ -1,0 +1,3 @@
+export const isAuthenticated = () => {
+  return !!localStorage.getItem("user");
+};
